Add knative to the demo
Add knative + contour to the demo, remove nginx and adapt
the ingress class to work with contour ingress controller.

This allows for all events to be send to a common broker and
then sent to the destination via knative triggers.
This setup allows to use something like cloudplayer to view
all the events regardless of the source.

Add a sequence diagram.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>